Build Database Script
Started writing a script to automate the database creation. Added SQL CREATE TABLE statements to their own python file as a list of strings.

connect() function needs review. I don't know what to put for user and password.

We should create the database in the project folder so we all have the same db state.

Co-Authored-By: ErrBrush <22796735+errbrush@users.noreply.github.com>